+++ CHANGELOG.md
@@ -7,7 +7,8 @@ since this pertains to binary (package) releases.
 
 ## [Package Unreleased]
 ### radar / 1090 Unreleased
-- Bump dependancy `ratatui` to `0.26.0```
+- Bump dependancy `ratatui` to `0.26.0`
+- Bump dependancy `deku` to `0.17.0`
 
 ## [2023-11-22]
 ### radar / 1090 v0.6.3